==> Source_Code/01_preprocessing.ipynb <==
import os
os.chdir("drive/My Drive/MLProject_20202")

import numpy as np
import pandas as pd
from bs4 import BeautifulSoup
from html import unescape
import unicodedata
import re
import warnings

data = pd.read_csv("Data/fake_job_postings.csv", delimiter=",")
pd.options.display.max_colwidth = None
data[11:12]["requirements"]

def beautify(s):
    s = str(s)
    s = BeautifulSoup(unescape(s), 'html').text
    s = unicodedata.normalize('NFKC', s)
    s = re.sub(r"([a-z:])([A-Z])", r"\1 \2", s)
    return s

s = beautify(float("NaN"))
s

df = pd.DataFrame(data = data)
warnings.filterwarnings("ignore", category=UserWarning, module='bs4')
for col in ["company_profile", "description", "requirements", "benefits"]:
    for i in range(len(df[col])):
        df.loc[i, col] = beautify(df.loc[i, col])
df["requirements"][11]

df.to_csv("Data/data_first_clean.csv", encoding='utf-8', index=False)
