==> Source_Code/02_data_overview.ipynb <==
import os
os.chdir('drive/My Drive/MLProject_20202')

import pandas as pd
import numpy as np

df = pd.read_csv('Data/data_first_clean.csv', encoding='utf-8')
df.head()

df = df.drop(columns=['job_id'])
df = df.fillna('')

df.describe(include='all')

# dropping categorical columns with too many variances or missing data
# some column like 'industry' are categorical but with too many categories
df = df.drop(columns=['industry', 'function', 'department', 'salary_range', 'required_education'])
df.head()

df.to_csv('Data/data_shortened.csv')
