==> Source_Code/03_one-hot-encoding.ipynb <==
import os
os.chdir('drive/My Drive/MLProject_20202')

import pandas as pd
import numpy as np

df = pd.read_csv('Data/data_shortened.csv', encoding='utf-8')
df.fillna('', inplace=True)
df.head()

from sklearn.preprocessing import OneHotEncoder

df.replace({'employment_type': {'': 'Other'}, 'required_experience': {'': 'Other', 'Not Applicable': 'Other'}}, inplace=True)

# one hot encoding these two columns
to_onehot = ['employment_type', 'required_experience']
for item in to_onehot:
    ohe = OneHotEncoder()
    ohe.fit(df[item].to_numpy().reshape(-1,1))
    transformed = ohe.transform(df[item].to_numpy().reshape(-1,1))
    ohe_df = pd.DataFrame.sparse.from_spmatrix(transformed)
    ohe_df.columns = ohe.get_feature_names([item])
    df = pd.concat([df, ohe_df], axis=1).drop([item], axis=1)

# this is cateogrical data only
df.iloc[:, 6:].astype(int).to_csv('Data/data_all_categorical.csv', encoding='utf-8', index=False)

# the first 6 columns are text data
df_text = df.iloc[:, :6]

df_text

df_text.to_csv('Data/data_all_text.csv', encoding='utf-8', index=False)
