==> Source_Code/04b_doc2vec.ipynb <==
import os
os.chdir('drive/My Drive/MLProject_20202')

!pip install --upgrade nltk

import pandas as pd
import pandas as pd
import numpy as np
import re

import nltk

nltk.download('punkt')
nltk.download('stopwords')

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize

stop_words = set(stopwords.words('english'))

df = pd.read_csv('Data/data_all_text.csv', encoding='utf-8')

df.fillna('', inplace=True)
df['concat_text'] = df['title'] + ' ' + df['location'] + ' ' + df['company_profile'] + ' ' + df['description'] + ' ' + df['requirements'] + ' ' + df['benefits']

def remove_punctuation(string: str):
    string = string.lower()
    apostrophe_replace = {"can't": "cannot","'s": " is", "'ve": " have", "'m": " am", "'re": " are", "n't": " not", "'d": " would", "'ll": " will"}
    for k,v in apostrophe_replace.items():
        string = re.sub(k, " " + v, string)
    
    string = re.sub(r" *-+", r" ", string)
    string = re.sub(r"-+ ", r" ", string)
    string = re.sub(r"#\w*", r"", string)
    string = re.sub(r'[^a-zA-Z0-9\-]', r' ', string)
    string = re.sub(r' +', r' ', string)

    return string

df['clean_text'] = df['concat_text'].apply(lambda x: remove_punctuation(x))

df['clean_text'].head()

df['clean_text'].to_csv("Data/data_text_clean.csv", encoding='utf-8', index=False)

documents = df.clean_text.tolist()

documents[:5]

from gensim.models.doc2vec import Doc2Vec, TaggedDocument

taggedDoc = [TaggedDocument(words=word_tokenize(_d.lower()), tags=[str(i)]) for i, _d in enumerate(documents)]

max_epochs = 100
vec_size = 200
alpha = 0.025

model = Doc2Vec(vector_size=vec_size,
                alpha=alpha, 
                min_alpha=0.00025,
                min_count=10,
                dm =1,
                workers=4)
  
model.build_vocab(taggedDoc)

for epoch in range(max_epochs):
    print('iteration {0}'.format(epoch))
    model.train(taggedDoc,
                total_examples=model.corpus_count,
                epochs=model.epochs)
    # decrease the learning rate
    model.alpha -= 0.0002
    # fix the learning rate, no decay
    model.min_alpha = model.alpha

model.save("Model/Doc2Vec/doc2vec_pv_dm_200.model")
print("Model Saved")

max_epochs = 100
vec_size = 200
alpha = 0.025

model = Doc2Vec(vector_size=vec_size,
                alpha=alpha, 
                min_alpha=0.00025,
                min_count=10,
                dm =0,
                workers=4)
  
model.build_vocab(taggedDoc)

for epoch in range(max_epochs):
    print('iteration {0}'.format(epoch))
    model.train(taggedDoc,
                total_examples=model.corpus_count,
                epochs=model.epochs)
    # decrease the learning rate
    model.alpha -= 0.0002
    # fix the learning rate, no decay
    model.min_alpha = model.alpha

model.save("Model/Doc2Vec/doc2vec_pv_dbow_200.model")
print("Model Saved")

max_epochs = 100
vec_size = 100
alpha = 0.025

model = Doc2Vec(vector_size=vec_size,
                alpha=alpha, 
                min_alpha=0.00025,
                min_count=10,
                dm =1,
                workers=4)
  
model.build_vocab(taggedDoc)

for epoch in range(max_epochs):
    print('iteration {0}'.format(epoch))
    model.train(taggedDoc,
                total_examples=model.corpus_count,
                epochs=model.epochs)
    # decrease the learning rate
    model.alpha -= 0.0002
    # fix the learning rate, no decay
    model.min_alpha = model.alpha

model.save("Model/Doc2Vec/doc2vec_pv_dm_100.model")
print("Model Saved")

max_epochs = 100
vec_size = 100
alpha = 0.025

model = Doc2Vec(vector_size=vec_size,
                alpha=alpha, 
                min_alpha=0.00025,
                min_count=10,
                dm =0,
                workers=4)
  
model.build_vocab(taggedDoc)

for epoch in range(max_epochs):
    print('iteration {0}'.format(epoch))
    model.train(taggedDoc,
                total_examples=model.corpus_count,
                epochs=model.epochs)
    # decrease the learning rate
    model.alpha -= 0.0002
    # fix the learning rate, no decay
    model.min_alpha = model.alpha

model.save("Model/Doc2Vec/doc2vec_pv_dbow_100.model")
print("Model Saved")
