==> Source_Code/04.5a_LSTM-text.ipynb <==
import os
os.chdir('drive/My Drive/MLProject_20202')

from keras import models

!pip install --upgrade keras

import json

with open('Data/dict_10.txt', 'r') as in_file:
    vocab = json.load(in_file)

with open('Data/reverse_dict_10.txt', 'r') as in_file:
    reverse_vocab = json.load(in_file)

import pandas as pd
import numpy as np
from keras.layers.merge import add
from keras.models import Sequential, Model
from keras.layers import Input, LSTM, Embedding, Dense,\
                        Activation, Dropout, Bidirectional
from keras.optimizers import Adam
from keras.utils.np_utils import to_categorical
from keras.preprocessing.sequence import pad_sequences
from keras.callbacks import Callback, ModelCheckpoint, EarlyStopping

#define length for feature and label
feature_max_length = 512
label_max_length = 8

df = pd.read_json('Data/data_text_embedded.json')
df.head()

def format_length(arr, length, vocab):
    while len(arr) < length:
        arr.append(vocab['<pad>'])
    newArr = arr[:(length-1)]
    if arr[length-1] == vocab['<pad>']:
        newArr.append(vocab['<pad>'])
    else:
        newArr.append(vocab['<end>'])
    return newArr

df.embedded_text = df.embedded_text.apply(lambda x: format_length(x, feature_max_length, vocab))
df.embedded_title = df.embedded_title.apply(lambda x: format_length(x, label_max_length, vocab))

df.embedded_title

# data generation for model
def data_generator(label, feature, max_length_title, batch_size, vocab_size):
    X1, X2, y = [], [], []
    n = 0
    # length of label/feature should be equal
    while 1:
        for i in range(len(label)):
            n += 1
            input = feature[i]
            output = label[i]
            for j in range(1,max_length_title):
                if output[j] == vocab['<end>']:
                    break
                else:
                    in_output = output[:j]
                    out_output = output[j]
                    in_output = pad_sequences([in_output], maxlen=max_length_title)[0]
                    out_output = to_categorical([out_output], num_classes=vocab_size)[0]
                    X1.append(input)
                    X2.append(in_output)
                    y.append(out_output)
            if n == batch_size:
                yield [np.array(X1), np.array(X2)], np.array(y)
                X1, X2, y = [], [], []
                n = 0

X = np.array(np.array(df['embedded_text']).tolist())
y = np.array(np.array(df['embedded_title']).tolist())

X

vocab_size = len(vocab)
embedding_dim = 256

inputs1 = Input(shape=(feature_max_length,))
fe1 = Embedding(vocab_size, embedding_dim, mask_zero=True)(inputs1)
fe2 = Dropout(0.5)(fe1)
# fe3 = LSTM(256)(fe2)
fe3 = Bidirectional(LSTM(256))(fe2)
inputs2 = Input(shape=(label_max_length,))
se1 = Embedding(vocab_size, embedding_dim, mask_zero=True)(inputs2)
se2 = Dropout(0.5)(se1)
# se3 = LSTM(256)(se2)
se3 = Bidirectional(LSTM(256))(se2)
decoder1 = add([fe3, se3])
decoder2 = Dense(256, activation='relu')(decoder1)
outputs = Dense(vocab_size, activation='softmax')(decoder2)
model = Model(inputs = [inputs1, inputs2], outputs = outputs)
model.summary()

model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy', 'MeanSquaredError'])

# model.optimizer.lr = 0.0001
model.optimizer.lr = 0.001
epochs = 100
batch_size = 128
steps = len(y)//batch_size
# checkpointer = ModelCheckpoint(filepath=os.path.join('/content/drive/My Drive/MLProject_20202/Model/LSTM/model_lstm_256_8.h5'), \
#                                monitor='val_loss', save_best_only=True, verbose=1)
checkpointer = ModelCheckpoint(filepath=os.path.join('/content/drive/My Drive/MLProject_20202/Model/LSTM/model_bilstm_512_8.h5'), \
                               monitor='val_loss', save_best_only=True, verbose=1)
es = EarlyStopping(monitor='val_loss', patience=5)

from sklearn.model_selection import KFold

num_fold = 5

kf = KFold(n_splits=num_fold, shuffle=True)
kf.get_n_splits(X)

for train_idx, test_idx in kf.split(X):
    X_train, X_test = X[train_idx], X[test_idx]
    y_train, y_test = y[train_idx], y[test_idx]
    train_generator = data_generator(y_train, X_train, label_max_length, batch_size, vocab_size)
    test_generator = data_generator(y_test, X_test, label_max_length, batch_size, vocab_size)
    model.fit(train_generator, epochs=epochs//num_fold, steps_per_epoch=len(X_train)//batch_size, \
              validation_data=test_generator, validation_steps=len(X_test)//batch_size,\
              callbacks=[checkpointer, es])
    print()
